#19. Made corresponding edits.
