+++ service/service/settings.py
@@ -56,7 +56,7 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    'custom_middleware.customlogger.LoggerMiddleware'
+    'middleware.logger.FiveGstLoggerMiddleware'
 ]
